Progress notifications in TopN classification
