+++ digits/inference/images/classification/__init__.py
@@ -1,4 +1,6 @@
 # Copyright (c) 2016, NVIDIA CORPORATION.  All rights reserved.
 from __future__ import absolute_import
 
+from .job import ImageInferenceClassificationJob
 from .job import ImageInferenceClassifyManyJob
+from .job import ImageInferenceTopNJob
