+++ digits/inference/images/classification/job.py
@@ -25,3 +25,30 @@ class ImageInferenceClassifyManyJob(ImageInferenceClassificationJob):
     @override
     def job_type(self):
         return 'Image Classification Classify Many Inference'
+
+@subclass
+class ImageInferenceTopNJob(ImageInferenceClassificationJob):
+    """
+    A Job that exercises the forward pass of a classification image neural network
+    Inference exercised through 'top N' method
+    """
+
+    @override
+    def __init__(self, **kwargs):
+        """
+        Keyword arguments:
+        top_n -- number of images to show in TopN view
+        """
+
+        self.top_n = kwargs.pop('top_n', None)
+
+        super(ImageInferenceTopNJob, self).__init__(**kwargs)
+
+    @override
+    def job_type(self):
+        return 'Image Classification TopN Inference'
+
+    @override
+    def get_parameters(self):
+        """Return inference parameters"""
+        return super(ImageInferenceTopNJob, self).get_parameters() + (self.top_n,)
